Moved h5_model to io.
